Refactor: simplify image import
