+++ lotti/lib/blocs/audio/recorder_cubit.dart
@@ -8,8 +8,8 @@ import 'package:lotti/classes/audio_note.dart';
 import 'package:lotti/classes/geolocation.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/insights_db.dart';
-import 'package:lotti/database/persistence_logic.dart';
 import 'package:lotti/location.dart';
+import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/main.dart';
 import 'package:lotti/utils/file_utils.dart';
 import 'package:lotti/utils/timezone.dart';
