+++ lotti/lib/main.dart
@@ -9,7 +9,6 @@ import 'package:intl/date_symbol_data_local.dart';
 import 'package:lotti/blocs/audio/player_cubit.dart';
 import 'package:lotti/blocs/audio/recorder_cubit.dart';
 import 'package:lotti/blocs/journal/health_cubit.dart';
-import 'package:lotti/blocs/journal/journal_image_cubit.dart';
 import 'package:lotti/blocs/sync/imap/inbox_cubit.dart';
 import 'package:lotti/blocs/sync/outbox_cubit.dart';
 import 'package:lotti/blocs/sync/sync_config_cubit.dart';
@@ -23,8 +22,8 @@ import 'package:lotti/widgets/home.dart';
 import 'package:window_manager/window_manager.dart';
 
 import 'database/database.dart';
-import 'database/persistence_logic.dart';
 import 'database/sync_db.dart';
+import 'logic/persistence_logic.dart';
 
 final getIt = GetIt.instance;
 
@@ -83,10 +82,6 @@ class LottiApp extends StatelessWidget {
         BlocProvider<HealthCubit>(
           lazy: true,
           create: (BuildContext context) => HealthCubit(),
-        ),
-        BlocProvider<JournalImageCubit>(
-          lazy: false,
-          create: (BuildContext context) => JournalImageCubit(),
         ),
         BlocProvider<AudioRecorderCubit>(
           create: (BuildContext context) => AudioRecorderCubit(),
