+++ lotti/lib/surveys/calculate.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:lotti/classes/journal_entities.dart';
-import 'package:lotti/database/persistence_logic.dart';
+import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/main.dart';
 import 'package:research_package/model.dart';
